+++ src/MixpostServiceProvider.php
@@ -14,6 +14,7 @@ use Inovector\Mixpost\Commands\ImportAccountData;
 use Inovector\Mixpost\Commands\ProcessMetrics;
 use Inovector\Mixpost\Commands\PruneTemporaryDirectory;
 use Inovector\Mixpost\Commands\PublishAssetsCommand;
+use Inovector\Mixpost\Commands\RefreshAccountTokens;
 use Inovector\Mixpost\Commands\RunScheduledPosts;
 use Inovector\Mixpost\Events\AccountAdded;
 use Inovector\Mixpost\Events\AccountUnauthorized;
@@ -53,6 +54,7 @@ class MixpostServiceProvider extends PackageServiceProvider
                 ProcessMetrics::class,
                 DeleteOldData::class,
                 PruneTemporaryDirectory::class,
+                RefreshAccountTokens::class,
             ])->hasInstallCommand(function (InstallCommand $command) {
                 $command
                     ->startWith(function (InstallCommand $command) {