+++ src/Schedule.php
@@ -14,5 +14,6 @@ class Schedule
         $schedule->command('mixpost:process-metrics')->everyThreeHours();
         $schedule->command('mixpost:delete-old-data')->daily();
         $schedule->command('mixpost:prune-temporary-directory')->hourly();
+        $schedule->command('mixpost:refresh-tokens')->daily();
     }
 }